pyproject.toml
[project]
name = "stock-watchlist-api"
version = "1.0.0"
description = "A multi-containerized stock watchlist REST API with FastAPI, MySQL, and Keycloak OAuth"
requires-python = ">=3.11"
dependencies = [
  "fastapi==0.104.1",
  "uvicorn[standard]==0.24.0",
  "sqlalchemy==2.0.23",
  "pymysql==1.1.0",
  "cryptography==41.0.7",
  "python-multipart==0.0.6",
  "pydantic==2.5.0",
  "google-auth==2.25.2",
  "requests==2.31.0",
  "aiohttp==3.9.1",
  "pyjwt[crypto]==2.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
dev-dependencies = []
